+++ src/vptstools/bin/vph5_to_vpts.py
@@ -1,22 +1,45 @@
+import os
 import tempfile
+from functools import partial
 import shutil
 from pathlib import Path
 from datetime import date
 
+import boto3
 import click
+from dotenv import load_dotenv
 import s3fs
 import pandas as pd
 
 from vptstools.vpts import vpts, vpts_to_csv
 from vptstools.s3 import handle_manifest, OdimFilePath
+from vptstools.bin.click_exception import catch_all_exceptions, report_exception_to_sns
 
-S3_BUCKET = "aloft"  # TODO - use config instead of hardcoded value
+# Load environmental variables from file in dev
+# (load_dotenv doesn't override existing environment variables)
+load_dotenv()
+
+S3_BUCKET = os.environ.get("DESTINATION_BUCKET", "aloft")
+INVENTORY_BUCKET = os.environ.get("INVENTORY_BUCKET", "aloft-inventory")
+AWS_SNS_TOPIC = os.environ.get("SNS_TOPIC")
+AWS_PROFILE = os.environ.get("AWS_PROFILE", None)
+AWS_REGION = os.environ.get("AWS_REGION", "eu-west-1")
+
+MANIFEST_URL = f"s3://{INVENTORY_BUCKET}/{S3_BUCKET}/{S3_BUCKET}-hdf5-files-inventory"
 S3_BUCKET_CREATION = pd.Timestamp("2022-08-02 00:00:00", tz="UTC")
-MANIFEST_URL = f"s3://aloft-inventory/{S3_BUCKET}/{S3_BUCKET}-hdf5-files-inventory"
 MANIFEST_HOUR_OF_DAY = "01-00"
 
 
-@click.command()
+# Prepare SNS report handler
+report_sns = partial(report_exception_to_sns,
+                     aws_sns_topic=AWS_SNS_TOPIC,
+                     subject="Conversion from hdf5 files to daily/monthly vpts-files failed.",
+                     profile_name=AWS_PROFILE,
+                     region_name=AWS_REGION
+                     )
+
+
+@click.command(cls=catch_all_exceptions(click.Command, handler=report_sns))  # Add SNS-reporting on exception
 @click.option(
     "--modified-days-ago",
     "modified_days_ago",
@@ -25,22 +48,16 @@ MANIFEST_HOUR_OF_DAY = "01-00"
     help="Range of h5 vp files to include, i.e. files modified between now and N"
     "modified-days-ago. If 0, all h5 files in the bucket will be included.",
 )
-@click.option(
-    "--aws-profile",
-    "aws_profile",
-    default=None,
-    help="(Optionally) AWS profile used to interact with the S3 bucket.",
-)
-def cli(modified_days_ago, aws_profile):
+def cli(modified_days_ago):
     """Convert and aggregate h5 vp files to daily/monthly vpts files on S3 bucket
 
     Check the latest modified h5 vp files on the S3 bucket using an S3 inventory,
     convert those files from ODIM bird profile to the VPTS CSV format and
     upload the generated daily/monthly vpts files to S3.
     """
-    if aws_profile:
-        storage_options = {"profile": aws_profile}
-        boto3_options = {"profile_name": aws_profile}
+    if AWS_PROFILE:
+        storage_options = {"profile": AWS_PROFILE}
+        boto3_options = {"profile_name": AWS_PROFILE}
     else:
         storage_options = dict()
         boto3_options = dict()
@@ -79,8 +96,6 @@ def cli(modified_days_ago, aws_profile):
     # Run vpts daily conversion for each radar-day with modified files
     inbo_s3 = s3fs.S3FileSystem(**storage_options)
     # PATCH TO OVERCOME RECURSIVE s3fs in wrapped context
-    import boto3
-
     session = boto3.Session(**boto3_options)
     s3_client = session.client("s3")