+++ CHANGELOG.md
@@ -2,4 +2,9 @@
 
 ## Version 0.1.0
 
-- hdf5 to csv base function with cli endpoint
+- Integrate functions from [odimh5](https://pypi.org/project/odimh5) to read odim5 files
+- Support for converting ODIm hdf5 files to the vpts-csv data standard
+- s3 data storage integration
+- CLI endpoint for the transfer of ODIM hdf5 files from Baltrad to the aloft S3 bucket
+- CLI endpoint for the conversion from ODIM hdf5 files to daily/monthly aggregates as vpts-csv format
+- Setup CI with Github Actions